+++ platform-sdk/consensus-event-creator-impl/src/main/java/org/hiero/consensus/event/creator/impl/tipset/TipsetEventCreator.java
@@ -1,15 +1,14 @@
 // SPDX-License-Identifier: Apache-2.0
-package com.swirlds.platform.event.creation.tipset;
+package org.hiero.consensus.event.creator.impl.tipset;
 
 import static com.swirlds.logging.legacy.LogMarker.EXCEPTION;
-import static com.swirlds.platform.event.creation.tipset.TipsetAdvancementWeight.ZERO_ADVANCEMENT_WEIGHT;
+import static org.hiero.consensus.event.creator.impl.tipset.TipsetAdvancementWeight.ZERO_ADVANCEMENT_WEIGHT;
 
 import com.hedera.hapi.node.base.SemanticVersion;
 import com.hedera.hapi.node.state.roster.Roster;
 import com.swirlds.base.time.Time;
 import com.swirlds.common.context.PlatformContext;
 import com.swirlds.common.utility.throttle.RateLimitedLogger;
-import com.swirlds.platform.event.EventUtils;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import edu.umd.cs.findbugs.annotations.Nullable;
 import java.time.Duration;
@@ -404,7 +403,7 @@ public class TipsetEventCreator implements EventCreator {
         if (lastSelfEvent == null) {
             timeCreated = now;
         } else {
-            timeCreated = EventUtils.calculateNewEventCreationTime(
+            timeCreated = calculateNewEventCreationTime(
                     now, lastSelfEvent.getTimeCreated(), lastSelfEvent.getTransactionCount());
         }
 
@@ -455,6 +454,33 @@ public class TipsetEventCreator implements EventCreator {
         }
 
         return sb.toString();
+    }
+
+    /**
+     * Calculate the creation time for a new event.
+     * <p>
+     * Regardless of whatever the host computer's clock says, the event creation time must always advance from self
+     * parent to child. Further, the time in between the self parent and the child must be large enough so that every
+     * transaction in the parent can be assigned a unique timestamp at nanosecond precision.
+     *
+     * @param now                        the current time
+     * @param selfParentCreationTime     the creation time of the self parent
+     * @param selfParentTransactionCount the number of transactions in the self parent
+     * @return the creation time for the new event
+     */
+    @NonNull
+    static Instant calculateNewEventCreationTime(
+            @NonNull final Instant now,
+            @NonNull final Instant selfParentCreationTime,
+            final int selfParentTransactionCount) {
+
+        final int minimumIncrement = Math.max(1, selfParentTransactionCount);
+        final Instant minimumNextEventTime = selfParentCreationTime.plusNanos(minimumIncrement);
+        if (now.isBefore(minimumNextEventTime)) {
+            return minimumNextEventTime;
+        } else {
+            return now;
+        }
     }
 
     /**
