+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/event/EventUtils.java
@@ -15,33 +15,6 @@ public final class EventUtils {
     private EventUtils() {}
 
     /**
-     * Calculate the creation time for a new event.
-     * <p>
-     * Regardless of whatever the host computer's clock says, the event creation time must always advance from self
-     * parent to child. Further, the time in between the self parent and the child must be large enough so that every
-     * transaction in the parent can be assigned a unique timestamp at nanosecond precision.
-     *
-     * @param now                        the current time
-     * @param selfParentCreationTime     the creation time of the self parent
-     * @param selfParentTransactionCount the number of transactions in the self parent
-     * @return the creation time for the new event
-     */
-    @NonNull
-    public static Instant calculateNewEventCreationTime(
-            @NonNull final Instant now,
-            @NonNull final Instant selfParentCreationTime,
-            final int selfParentTransactionCount) {
-
-        final int minimumIncrement = Math.max(1, selfParentTransactionCount);
-        final Instant minimumNextEventTime = selfParentCreationTime.plusNanos(minimumIncrement);
-        if (now.isBefore(minimumNextEventTime)) {
-            return minimumNextEventTime;
-        } else {
-            return now;
-        }
-    }
-
-    /**
      * Returns the timestamp of the last transaction in this event. If this event has no transaction, then the timestamp
      * of the event will be returned
      *
